+++ dispositif_ajout.sql
@@ -54,14 +54,14 @@ select
     'Ajouter' as title,
     'dispositif_ajout.sql' as link,
     'square-rounded-plus' as icon,
-        $group_id::int<3 as disabled,
+        $group_id<3 as disabled,
     'green' as outline;
     */
 select 
     'Supprimer' as title,
     'dispositif_delete.sql' as link,
     'trash' as icon,
-        $group_id::int<3 as disabled,
+        $group_id<3 as disabled,
     'red' as outline;
 
 -- Liste et ajout
