+++ etab_classes_boutons.sql
@@ -163,7 +163,7 @@ THEN
 ](notification.sql?id='||eleve.id||'&tab=Examen)' 
 ELSE ''
 END as Actions,
-CASE WHEN $group_id::int>2 THEN
+CASE WHEN $group_id>2 THEN
 '[
     ![](./icons/pencil.svg)
 ](eleve_edit.sql?id='||eleve.id||')'
